Enormous entity restructuring, serialization changes, and miscellaneous culling.
First off, I'm doing something that I think is extremely bizarre and strange in C++. Instead of relying on variables to depict properties for game entities (ie an item's name, damage values, etc.) I'm relying on a structure similar to this:

std::map< std::string, boost::any >

This allows for statements such as this to compile into valid c++

_properties["name"] = "sword";   // std::string
_properties["damage"] = 50;      // int
_properties["speed"] = 25.3f;    // float

All entities contain a _properties value map which is directly serialized instead of hard-coded serialization with AMEF protocol. Serialization is now XML-based, which might be migrated to JSON or something else should I find it necessary. For now it's plaintext and human-readable, which I quite like.

Also added the beginning of the World module, which is the global container of Map types.